Re-organization of the benchmark code
